fix(components/toast): move harness files under `/toast` to match docs (#3173)
