+++ libs/components/toast/testing/src/public-api.ts
@@ -1,5 +1,5 @@
 export { SkyToastFixture } from './legacy/toast-fixture';
 
-export { SkyToasterHarness } from './modules/toaster-harness';
-export { SkyToastHarness } from './modules/toast-harness';
-export { SkyToastHarnessFilters } from './modules/toast-harness-filters';
+export { SkyToasterHarness } from './modules/toast/toaster-harness';
+export { SkyToastHarness } from './modules/toast/toast-harness';
+export { SkyToastHarnessFilters } from './modules/toast/toast-harness-filters';
